Windows support: fail to configure the `k8s` plugin
